Apply the same -f formatting changes to translate.
