+++ translate.py
@@ -18,8 +18,85 @@ import os
 import sys
 import click
 import whisper
+import json
 
 from mlhub.pkg import get_cmd_cwd
+
+# -----------------------------------------------------------------------
+# Helper functions
+# Copied and modified based on functions from OpenAI's whisper package,
+# https://github.com/openai/whisper/blob/main/whisper/utils.py
+# -----------------------------------------------------------------------
+
+# Utility function to format timestamps
+def format_timestamp(seconds: float, format_type: str = "srt"):
+    milliseconds = round(seconds * 1000.0)
+
+    hours = milliseconds // 3_600_000
+    milliseconds -= hours * 3_600_000
+
+    minutes = milliseconds // 60_000
+    milliseconds -= minutes * 60_000
+
+    seconds = milliseconds // 1_000
+    milliseconds -= seconds * 1_000
+
+    if format_type == "srt":
+        return f"{int(hours):02}:{int(minutes):02}:{int(seconds):02},{milliseconds:03d}"
+    elif format_type == "vtt":
+        hours_marker = f"{hours:02d}:" if hours > 0 else ""
+        return f"{hours_marker}{int(minutes):02}:{int(seconds):02}.{milliseconds:03d}"
+
+# Output Handler Class
+class OutputHandler:
+    def __init__(self, format, output_path=None):
+        self.format = format
+        self.output_path = output_path
+
+    def _output_txt(self, result, file):
+        for segment in result["segments"]:
+            print(segment["text"].strip(), file=file, flush=True)
+
+    def _output_json(self, result, file):
+        json.dump(result, file, indent=2)
+
+    def _output_srt(self, result, file):
+        for i, segment in enumerate(result["segments"], start=1):
+            start = format_timestamp(segment["start"], "srt")
+            end = format_timestamp(segment["end"], "srt")
+            print(f"{i}\n{start} --> {end}\n{segment['text'].strip()}\n", file=file, flush=True)
+
+    def _output_vtt(self, result, file):
+        print("WEBVTT\n", file=file)
+        for segment in result["segments"]:
+            start = format_timestamp(segment["start"], "vtt")
+            end = format_timestamp(segment["end"], "vtt")
+            print(f"{start} --> {end}\n{segment['text'].strip()}\n", file=file, flush=True)
+
+    def _output_tsv(self, result, file):
+        print("start", "end", "text", sep="\t", file=file)
+        for segment in result["segments"]:
+            print(round(1000 * segment["start"]), file=file, end="\t")
+            print(round(1000 * segment["end"]), file=file, end="\t")
+            print(segment["text"].strip().replace("\t", " "), file=file, flush=True)
+
+    def write(self, result):
+        # Dynamically select the appropriate output method.
+        # If a method for the specified format does not exist, default to txt.
+        output_func = getattr(self, f"_output_{self.format}", self._output_txt)
+
+        if self.output_path:
+            if os.path.exists(self.output_path):
+                # File exists, write an error message and exit with status 1
+                print(f"Error: File '{self.output_path}' already exists.", file=sys.stderr)
+                sys.exit(1)
+            else: # File does not exist
+                with open(self.output_path, "w", encoding="utf-8") as f:
+                    output_func(result, f)
+                print(f"Transcribed text saved to {self.output_path}", file=sys.stderr)
+                sys.exit(0)
+        else:
+            output_func(result, sys.stdout)
 
 # -----------------------------------------------------------------------
 # Command line argument and options
@@ -37,11 +114,11 @@ from mlhub.pkg import get_cmd_cwd
 @click.option("-f", "--format",
               default=None,
               type=click.STRING,
-              help="The format of the output file. e.g. txt, json, srt")
+              help="The format of the output. Supported formats are txt, json, srt, tsv, and vtt.")
 @click.option("-o", "--output",
               default=None,
               type=click.STRING,
-              help="The name and format of the output file. e.g. output.txt")
+              help="The name and format of the output file. e.g. output.txt, tmp.vtt")
 
 def cli(filename, lang, output, format):
     """
@@ -54,11 +131,13 @@ def cli(filename, lang, output, format):
 
     Use the `-l` or `--lang` option to specify the language of the source audio.
 
-    To save the translated text to a file, 
-    use the `-o` or `--output` option to specify the desired output file name and 
-    format (e.g. `output.txt`), 
-    or use the `-f` or `--format` option to specify the desired output file format
-    (e.g. `txt`) while the file name will be the same as input audio file name.
+    Use the `-f` or `--format` option to specify the desired output format. 
+    Supported formats are txt, json, srt, tsv, and vtt.
+    (e.g. `-f txt`).
+
+    To save the transcribed text to a file, 
+    Use the `-o` or `--output` option to specify the desired output file name 
+    and format (e.g. `-o output.txt`),
 
     """
 
@@ -85,42 +164,29 @@ def cli(filename, lang, output, format):
         
     result = model.transcribe(path, fp16=False, task="translate", language=lang)
 
-    if output or format:
-        output_path = (
-            os.path.join(get_cmd_cwd(), output) if output 
-            else os.path.join(get_cmd_cwd(), 
-                              filename.replace(filename.split(".")[-1], format))
-        )
-
-    text_buffer = [] # Buffer for accumulating segments of one sentence.
-
-    # Process and output the text ensuring one sentence per line.
-    for segment in result["segments"]:
-        text_buffer.append(segment["text"].strip())
+    if format or output:
+        output_format = format if format else output.split(".")[-1]
+        output_path = os.path.join(get_cmd_cwd(), output) if output else None
+        output_handler = OutputHandler(output_format, output_path)
+        output_handler.write(result)
+    else: 
+        text_buffer = [] # Buffer for accumulating segments of one sentence.
         
-        if segment["text"].strip()[-1] in [".", "?", "!", "。", "？", "！"]:
-            # Reached the end of a sentence.
-            full_sentence = " ".join(text_buffer)
+        # If no format or output is specified, 
+        # print the text to the console as one sentence per line.
+        for segment in result["segments"]:
+            text_buffer.append(segment["text"].strip())
             
-            if output or format:
-                with open(output_path, "a", encoding="utf-8") as f:
-                    f.write(full_sentence + "\n")
-            else:
+            if segment["text"].strip()[-1] in [".", "?", "!", "。", "？", "！"]:
+                # Reached the end of a sentence.
+                full_sentence = " ".join(text_buffer)
                 print(full_sentence)
-            
-            text_buffer = []
-    
-    # Handle the remaining text in the buffer.
-    if text_buffer:
-        trailing_text = " ".join(text_buffer)
-        if output or format:
-            with open(output_path, "a", encoding="utf-8") as f:
-                f.write(trailing_text + "\n")
-        else:
+                text_buffer = []
+        
+        # Handle the remaining text in the buffer.
+        if text_buffer:
+            trailing_text = " ".join(text_buffer)
             print(trailing_text)
-
-    if output or format:
-        print("Translated text saved to", output_path)
     
 if __name__ == "__main__":
     cli(prog_name="translate")
